added cookie feature for remembering logged in user

+++ src/main/java/com/expense/trackig/ExpenseTracking/Controllers/CategoryController.java
@@ -2,6 +2,8 @@ package com.expense.trackig.ExpenseTracking.Controllers;
 
 import com.expense.trackig.ExpenseTracking.Modules.Category;
 import com.expense.trackig.ExpenseTracking.Services.Implementation.CategoryServiceImpl;
+import com.expense.trackig.ExpenseTracking.Services.Interface.CategoriesService;
+import com.expense.trackig.ExpenseTracking.Services.Interface.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -11,16 +13,22 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/category")
 public class CategoryController {
     @Autowired
-    CategoryServiceImpl categoryService;
+    CategoriesService categoryService;
+    @Autowired
+    UserService userService;
 
     @GetMapping("/dashboard")
-    public String dashboard(Model model){
+    public String dashboard(@CookieValue(value = "expense_cookie", defaultValue = "null") String cookie,Model model){
+        if(!userService.isLoggedIn(cookie))
+            return "redirect:/public/login";
         model.addAttribute("categories",categoryService.getAllCategories());
         return "dashboard";
     }
 
     @PostMapping("/add")
-    public String addCategory(@ModelAttribute Category category){
+    public String addCategory(@CookieValue(value = "expense_cookie", defaultValue = "null") String cookie,@ModelAttribute Category category){
+        if(!userService.isLoggedIn(cookie))
+            return "redirect:/public/login";
         try {
             categoryService.saveCategory(category);
         }catch (Exception e){
@@ -29,7 +37,9 @@ public class CategoryController {
         return "redirect:/category/dashboard";
     }
     @GetMapping("/remove/{name}")
-    public String removeCategory(@PathVariable("name") String name){
+    public String removeCategory(@CookieValue(value = "expense_cookie", defaultValue = "null") String cookie,@PathVariable("name") String name){
+        if(!userService.isLoggedIn(cookie))
+            return "redirect:/public/login";
         try {
             categoryService.removeCategory(name);
         }catch (Exception e){
